+++ mobile/android/android-components/buildSrc/src/main/java/Dependencies.kt
@@ -132,6 +132,8 @@ object Dependencies {
 
     const val mozilla_push = "org.mozilla.appservices:push:${Versions.mozilla_appservices}"
 
+    const val mozilla_remote_tabs = "org.mozilla.appservices:tabs:${Versions.mozilla_appservices}"
+
     const val mozilla_httpconfig = "org.mozilla.appservices:httpconfig:${Versions.mozilla_appservices}"
     const val mozilla_full_megazord = "org.mozilla.appservices:full-megazord:${Versions.mozilla_appservices}"
     const val mozilla_full_megazord_forUnitTests = "org.mozilla.appservices:full-megazord-forUnitTests:${Versions.mozilla_appservices}"